Pull out duplicate data gen
